requirements.txt
backports-abc==0.5
certifi==2017.7.27.1
click==6.7
Django==1.10
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Jinja2==2.9.6
livereload==2.5.1
Markdown==2.6.9
MarkupSafe==1.0
mkdocs==0.16.3
psycopg2==2.7.3.1
PyJWT==1.5.2
python-decouple==3.1
PyYAML==3.12
singledispatch==3.4.0.3
six==1.10.0
tornado==4.5.1
